tests/wpt/tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.drawing.style.fontKerning.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.text.drawing.style.fontKerning</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.fontKerning</h1>
<p class="desc">Testing basic functionalities of fontKerning for canvas</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing basic functionalities of fontKerning for canvas");
_addTest(function(canvas, ctx) {

  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "normal";
  _assertSame(ctx.fontKerning, "normal", "ctx.fontKerning", "\"normal\"");
  width_normal = ctx.measureText("TAWATAVA").width;
  ctx.fontKerning = "none";
  _assertSame(ctx.fontKerning, "none", "ctx.fontKerning", "\"none\"");
  width_none = ctx.measureText("TAWATAVA").width;
  _assert(width_normal < width_none, "width_normal < width_none");

});
</script>
